+++ Sources/Utils/GUIUtils/LocalizableItemAttribute.cs
@@ -9,10 +9,10 @@ namespace KSPDev.GUIUtils {
 /// <summary>Attribute for the various game items that support localization.</summary>
 /// <remarks>
 /// <para>
-/// This attribute alone doesn't make the annotated item localized. There should be a code executed
-/// that understands the type of the item. However, it still makes sense to add this attribute to
-/// <i>any</i> game item that needs localization, it will help the <c>LocalizationTool</c> to
-/// extract the information.
+/// Use this attribute to automatically localize the part's events, fields, and actions. The
+/// relevant strings will be loaded when the vessel or part is either created or loaded. The
+/// strings will also properly update when reloading the language definitiaion via
+/// <c>LocalizationTool</c>.
 /// </para>
 /// <para>
 /// See the "seealso" section for the types and methods that are aware of this attribute.